+++ frontend/src/components/owner/owner_applications/OwnerRequestsTable.tsx
