+++ frontend/src/components/owner/owner_profile/OwnerProfilePlaceholder.tsx
@@ -1,0 +1,14 @@
+import { useTranslation } from 'react-i18next'
+import styles from '@/styles/OwnerDashboard.module.css'
+
+export default function OwnerProfilePlaceholder() {
+  const { t } = useTranslation()
+
+  return (
+    <section className={styles.ownerPlaceholder}>
+      <span className={styles.ownerPlaceholderKicker}>{t('ownerDashboard.placeholder.kicker')}</span>
+      <h1 className={styles.ownerPlaceholderTitle}>{t('ownerDashboard.sidebar.profile')}</h1>
+      <p className={styles.ownerPlaceholderSubtitle}>{t('ownerDashboard.placeholder.subtitle')}</p>
+    </section>
+  )
+}
