+++ frontend/src/components/owner/owner_properties/OwnerSummaryCard.tsx
