+++ frontend/src/pages/owner/OwnerMessages.tsx
@@ -1,17 +1,10 @@
-import { useTranslation } from 'react-i18next'
-import OwnerDashboardLayout from '@/components/owner/OwnerDashboardLayout'
-import styles from '@/styles/OwnerDashboard.module.css'
+import OwnerLayout from '@/components/owner/OwnerLayout'
+import OwnerMessagesPlaceholder from '@/components/owner/owner_messages/OwnerMessagesPlaceholder'
 
 export default function OwnerMessages() {
-  const { t } = useTranslation()
-
   return (
-    <OwnerDashboardLayout>
-      <section className={styles.ownerPlaceholder}>
-        <span className={styles.ownerPlaceholderKicker}>{t('ownerDashboard.placeholder.kicker')}</span>
-        <h1 className={styles.ownerPlaceholderTitle}>{t('ownerDashboard.sidebar.messages')}</h1>
-        <p className={styles.ownerPlaceholderSubtitle}>{t('ownerDashboard.placeholder.subtitle')}</p>
-      </section>
-    </OwnerDashboardLayout>
+    <OwnerLayout>
+      <OwnerMessagesPlaceholder />
+    </OwnerLayout>
   )
 }
