+++ frontend/src/pages/owner/OwnerNotifications.tsx
@@ -1,17 +1,10 @@
-import { useTranslation } from 'react-i18next'
-import OwnerDashboardLayout from '@/components/owner/OwnerDashboardLayout'
-import styles from '@/styles/OwnerDashboard.module.css'
+import OwnerLayout from '@/components/owner/OwnerLayout'
+import OwnerNotificationsPlaceholder from '@/components/owner/owner_notifications/OwnerNotificationsPlaceholder'
 
 export default function OwnerNotifications() {
-  const { t } = useTranslation()
-
   return (
-    <OwnerDashboardLayout>
-      <section className={styles.ownerPlaceholder}>
-        <span className={styles.ownerPlaceholderKicker}>{t('ownerDashboard.placeholder.kicker')}</span>
-        <h1 className={styles.ownerPlaceholderTitle}>{t('ownerDashboard.sidebar.notifications')}</h1>
-        <p className={styles.ownerPlaceholderSubtitle}>{t('ownerDashboard.placeholder.subtitle')}</p>
-      </section>
-    </OwnerDashboardLayout>
+    <OwnerLayout>
+      <OwnerNotificationsPlaceholder />
+    </OwnerLayout>
   )
 }
