+++ frontend/src/pages/owner/OwnerPublishPropertyPage.tsx
@@ -1,8 +1,8 @@
 import { ArrowLeftIcon } from '@heroicons/react/24/outline'
 import { useTranslation } from 'react-i18next'
 import { Link } from 'react-router-dom'
-import OwnerDashboardLayout from '@/components/owner/OwnerDashboardLayout'
-import OwnerPropertyPublishForm from '@/components/owner/OwnerPropertyPublishForm'
+import OwnerLayout from '@/components/owner/OwnerLayout'
+import OwnerPropertyPublishForm from '@/components/owner/owner_publish_property/OwnerPropertyPublishForm'
 import { paths } from '@/routes/paths'
 import styles from '@/styles/OwnerPublishProperty.module.css'
 
@@ -10,12 +10,12 @@ export default function OwnerPublishPropertyPage() {
   const { t } = useTranslation()
 
   return (
-    <OwnerDashboardLayout>
+    <OwnerLayout>
       <Link to={paths.ownerProperties} className={styles.backLink}>
         <ArrowLeftIcon className={styles.backIcon} aria-hidden="true" />
         {t('ownerDashboard.publish.backToProperties')}
       </Link>
       <OwnerPropertyPublishForm />
-    </OwnerDashboardLayout>
+    </OwnerLayout>
   )
 }
